+++ doc/source/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 
 X.Y.Z (DD-MM-YYYY)
 ------------------
+* Test xarray-ms against the full MSv4 Test Corpus (:pr:`120`)
 * Cleanup Dataset Factories (:pr:`118`)
 * Add field_and_source_xds (:pr:`114`, :pr:`119`, :pr:`122`)
 * Fix integer selection (:pr:`117`)